requirements.txt
APScheduler           == 3.6.3
backports.zoneinfo    == 0.2.1
cachetools            == 4.2.2
certifi               == 2021.10.8
cycler                == 0.11.0
fonttools             == 4.29.1
grpcio                == 1.44.0
grpcio-tools          == 1.44.0
kiwisolver            == 1.3.2
matplotlib            == 3.5.1
numpy                 == 1.22.2
packaging             == 21.3
Pillow                == 9.0.1
pip                   == 22.0.3
protobuf              == 3.19.4
pyparsing             == 3.0.7
python-dateutil       == 2.8.2
python-dotenv         == 0.19.1
python-telegram-bot   == 13.11
pytz                  == 2021.3
pytz-deprecation-shim == 0.1.0.post0
setuptools            == 44.0.0
six                   == 1.16.0
tornado               == 6.1
tzdata                == 2021.5
tzlocal               == 4.1
wheel                 == 0.36.2
